+++ src/StructuredField.php
@@ -4,10 +4,12 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
+use DateTimeInterface;
 use Stringable;
 
 /**
- * @phpstan-type SfTypeInput StructuredField|ByteSequence|Token|\DateTimeInterface|string|int|float|bool
+ * @phpstan-type SfType ByteSequence|Token|DateTimeInterface|string|int|float|bool
+ * @phpstan-type SfTypeInput StructuredField|SfType
  * @phpstan-type SfItem ValueAccess&ParameterAccess
  * @phpstan-type SfItemInput SfItem|SfTypeInput
  * @phpstan-type SfMember (MemberList<int, SfItem>|ValueAccess)&ParameterAccess